requirements.txt
pandas>=2.2
pyarrow>=16.0
requests>=2.32
python-dotenv>=1.0
matplotlib>=3.9
